+++ app/views/main.scala.html
@@ -50,11 +50,6 @@
     </div>
 
     @content
-    <div>
-        <button class="w3-button w3-white w3-large" onclick="sideOpen()">&#9776;</button>
-    </div>
-
-        @content
     <br>
     <div class="footer" id="myFooter">
         <div class="row">
@@ -67,13 +62,13 @@
 
 
 
+    <script src="@routes.Assets.versioned("javascripts/hello.js")" type="text/javascript"></script>
     <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js"
     integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n"
     crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/js/bootstrap.min.js"
     integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6"
     crossorigin="anonymous"></script>
-    <script src="@routes.Assets.versioned("javascripts/hello.js")" type="text/javascript"></script>
 
     </body>
 </html>
