+++ public/stylesheets/main.css
@@ -209,17 +209,3 @@ p{
     color: black;
     font-size: 40px;
 }
-.carousel-inner .carousel-item-right.active,
-.carousel-inner .carousel-item-next {
-    transform: translateX(33.33%);
-}
-
-.carousel-inner .carousel-item-left.active,
-.carousel-inner .carousel-item-prev {
-    transform: translateX(-33.33%)
-}
-
-.carousel-inner .carousel-item-right,
-.carousel-inner .carousel-item-left{
-    transform: translateX(0);
-}
